+++ builds/itee-plugin-three.esm.js
@@ -1,4 +1,4 @@
-console.log('Itee.Plugin.Three v1.2.4 - EsModule')
+console.log('Itee.Plugin.Three v1.2.7 - EsModule')
 import { DefaultLogger, TBinaryReader, Endianness, FileFormat, Keys, Mouse, TDataBaseManager, ResponseType } from 'itee-client';
 import { DefaultLoadingManager, Box3, FileLoader, Group, BufferGeometry, BufferAttribute, PointsMaterial, Points, TextureLoader, Mesh, PlaneGeometry, MeshBasicMaterial, DoubleSide, MathUtils, Vector3, Shape, ColladaLoader, FBXLoader, ObjectLoader, OBJLoader, ShapeBufferGeometry, MeshPhongMaterial, STLLoader, MTLLoader, EventDispatcher, Object3D, Camera, Vector2, Spherical, Quaternion, MOUSE, LineSegments, EdgesGeometry, BoxBufferGeometry, LineBasicMaterial, Plane, CylinderBufferGeometry, Float32BufferAttribute, OctahedronBufferGeometry, SphereBufferGeometry, TorusBufferGeometry, Line, ConeBufferGeometry, PlaneBufferGeometry, Raycaster, Euler, Scene, SplineCurve, QuadraticBezierCurve3, QuadraticBezierCurve, Path, LineCurve3, LineCurve, EllipseCurve, CurvePath, Curve, CubicBezierCurve3, CubicBezierCurve, CatmullRomCurve3, ArcCurve, WireframeGeometry, SphereGeometry, TubeGeometry, TorusKnotGeometry, TorusGeometry, TextGeometry, TetrahedronGeometry, ShapeGeometry, RingGeometry, PolyhedronGeometry, ParametricGeometry, OctahedronGeometry, LatheGeometry, IcosahedronGeometry, Geometry, ExtrudeGeometry, DodecahedronGeometry, ConeGeometry, CylinderGeometry, CircleGeometry, BoxGeometry, Face3, InstancedBufferGeometry, TubeBufferGeometry, TorusKnotBufferGeometry, TextBufferGeometry, TetrahedronBufferGeometry, RingBufferGeometry, PolyhedronBufferGeometry, ParametricBufferGeometry, LatheBufferGeometry, IcosahedronBufferGeometry, ExtrudeBufferGeometry, DodecahedronBufferGeometry, CircleBufferGeometry, ImageLoader, MeshLambertMaterial, Color, LinearFilter, Sprite, LineLoop, LOD, SkinnedMesh, HemisphereLight, SpotLight, RectAreaLight, PointLight, DirectionalLight, AmbientLight, OrthographicCamera, PerspectiveCamera, Fog, FogExp2, VertexColors } from 'three-full';
 import { degreesToRadians, toEnum } from 'itee-utils';
@@ -5969,6 +5969,9 @@ class TranslateHandle extends AbstractHandle {
             this.flipDirection();
         }
 
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
+
     }
 
     setDirection ( direction ) {
@@ -6061,6 +6064,9 @@ class ScaleHandle extends AbstractHandle {
             this.flipDirection();
         }
 
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
+
     }
 
     setDirection ( direction ) {
@@ -6094,6 +6100,10 @@ class RotateHandle extends AbstractHandle {
 
     update ( cameraDirection ) {
         super.update( cameraDirection );
+
+
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
     }
 
 }
@@ -6284,6 +6294,9 @@ class PlaneHandle extends AbstractHandle {
 
         }
 
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
+
     }
 
     setDirection ( direction ) {
@@ -6457,6 +6470,9 @@ class LozengeHandle extends AbstractHandle {
 
         }
 
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
+
     }
 
     setDirection ( direction ) {
@@ -6510,7 +6526,7 @@ class OctahedricalHandle extends AbstractHandle {
         this.isOmnidirectionalHandle = true;
         this.type                    = 'OmnidirectionalHandle';
 
-        const octahedronGeometry    = new OctahedronBufferGeometry( 1, 0 );
+        const octahedronGeometry    = new OctahedronBufferGeometry( 0.1, 0 );
         const octahedronMaterial    = new HighlightableMaterial( {
             color:       _parameters.color,
             transparent: true,
@@ -6533,6 +6549,9 @@ class OctahedricalHandle extends AbstractHandle {
 
     update ( cameraDirection ) {
         super.update( cameraDirection );
+
+        this.updateMatrix();
+        this.hitbox.updateMatrix();
     }
 
 }
@@ -6567,7 +6586,7 @@ class AbstractGizmo extends Object3D {
         } );
         this.intersectPlane                  = new Mesh( planeGeometry, planeMaterial );
         this.intersectPlane.matrixAutoUpdate = false;
-        this.intersectPlane.visible          = false;
+        this.intersectPlane.visible          = true;
 
         this.add( this.intersectPlane );
 
@@ -6668,6 +6687,7 @@ class AbstractGizmo extends Object3D {
     updateIntersectPlane ( cameraPosition ) {
 
         this.intersectPlane.lookAt( cameraPosition );
+        this.intersectPlane.updateMatrix();
 
     }
 
@@ -7126,6 +7146,8 @@ class ClippingControls extends Object3D {
 
         }
 
+        this.updateGizmo();
+
     }
 
     setCamera ( value ) {
@@ -7248,10 +7270,12 @@ class ClippingControls extends Object3D {
             this.position.set( this._objectsToClipCenter.x, this._objectsToClipCenter.y, this._objectsToClipCenter.z );
 
             // update...
+            this.updateMatrix();
             this.updateMatrixWorld();
         }
 
         this.updateClipping();
+        this.updateGizmo();
 
     }
 
@@ -7272,7 +7296,7 @@ class ClippingControls extends Object3D {
 
     }
 
-    update () {
+    updateGizmo () {
 
         if ( !this.enabled ) { return }
         if ( this._mode === ClippingModes.None ) { return }
@@ -7709,6 +7733,8 @@ class ClippingControls extends Object3D {
 
         }
 
+        this.updateGizmo();
+
     }
 
     _onMouseWheel ( mouseEvent ) {
@@ -7803,24 +7829,28 @@ class ClippingControls extends Object3D {
     _translate ( displacement ) {
 
         this.position.add( displacement );
+        this.updateMatrix();
 
     }
 
     _translateX ( deltaX ) {
 
         this.position.setX( this.position.x + deltaX );
+        this.updateMatrix();
 
     }
 
     _translateY ( deltaY ) {
 
         this.position.setY( this.position.y + deltaY );
+        this.updateMatrix();
 
     }
 
     _translateZ ( deltaZ ) {
 
         this.position.setZ( this.position.z + deltaZ );
+        this.updateMatrix();
 
     }
 
@@ -7828,6 +7858,7 @@ class ClippingControls extends Object3D {
 
         this.position.setX( this.position.x + deltaX );
         this.position.setY( this.position.y + deltaY );
+        this.updateMatrix();
 
     }
 
@@ -7835,6 +7866,7 @@ class ClippingControls extends Object3D {
 
         this.position.setX( this.position.x + deltaX );
         this.position.setZ( this.position.z + deltaZ );
+        this.updateMatrix();
 
     }
 
@@ -7842,12 +7874,14 @@ class ClippingControls extends Object3D {
 
         this.position.setY( this.position.y + deltaY );
         this.position.setZ( this.position.z + deltaZ );
+        this.updateMatrix();
 
     }
 
     _translateXYZ ( deltaX, deltaY, deltaZ ) {
 
         this.position.set( this.position.x + deltaX, this.position.y + deltaY, this.position.z + deltaZ );
+        this.updateMatrix();
 
     }
 
@@ -7877,24 +7911,28 @@ class ClippingControls extends Object3D {
     _scale ( changeAmout ) {
 
         this.scale.add( changeAmout );
+        this.updateMatrix();
 
     }
 
     _scaleX ( deltaX ) {
 
         this.scale.setX( this.scale.x + deltaX );
+        this.updateMatrix();
 
     }
 
     _scaleY ( deltaY ) {
 
         this.scale.setY( this.scale.y + deltaY );
+        this.updateMatrix();
 
     }
 
     _scaleZ ( deltaZ ) {
 
         this.scale.setZ( this.scale.z + deltaZ );
+        this.updateMatrix();
 
     }
 
@@ -7902,6 +7940,7 @@ class ClippingControls extends Object3D {
 
         this.scale.setX( this.scale.x + deltaX );
         this.scale.setY( this.scale.y + deltaY );
+        this.updateMatrix();
 
     }
 
@@ -7909,6 +7948,7 @@ class ClippingControls extends Object3D {
 
         this.scale.setX( this.scale.x + deltaX );
         this.scale.setZ( this.scale.z + deltaZ );
+        this.updateMatrix();
 
     }
 
@@ -7916,12 +7956,14 @@ class ClippingControls extends Object3D {
 
         this.scale.setY( this.scale.y + deltaY );
         this.scale.setZ( this.scale.z + deltaZ );
+        this.updateMatrix();
 
     }
 
     _scaleXYZ ( deltaX, deltaY, deltaZ ) {
 
         this.scale.set( this.scale.x + deltaX, this.scale.y + deltaY, this.scale.z + deltaZ );
+        this.updateMatrix();
 
     }